refactor: new ui with job & header
