+++ console/src/components/Card/index.module.scss
@@ -15,7 +15,7 @@
     transition: transform 270ms cubic-bezier(0.1, 0.9, 0.2, 1),opacity 300ms cubic-bezier(0.1, 0.9, 0.2, 1) 0.05s,-webkit-transform 300ms cubic-bezier(0.1, 0.9, 0.2, 1);
     transform: translate3d(0,48px,0);
     transition-delay: 0.4s;
-    margin-bottom: 16px;
+    margin-bottom: 10px;
     display: flex;
     border-radius: 8px;
     padding: 36px 28px;
