+++ console/src/components/Header/index.tsx
@@ -18,17 +18,19 @@ import User from '@/domain/user/components/User'
 import { simulationJump } from '@/utils'
 import { FiLogOut, FiUser } from 'react-icons/fi'
 import { BsChevronDown } from 'react-icons/bs'
+import { StatefulMenu } from 'baseui/menu'
 
 const useHeaderStyles = createUseStyles({
     headerWrapper: {
         padding: '0 32px 0 0',
-        // position: 'fixed',
+        position: 'absolute',
         // background: 'var(--color-brandHeaderBackground)',
-        backdropFilter: 'blur(10px)',
+        // backdropFilter: 'blur(10px)',
         zIndex: 1000,
-        // top: 0,
+        top: 0,
+        right: 0,
         height: `${headerHeight}px`,
-        width: '100%',
+        width: '30%',
         display: 'flex',
         flexFlow: 'row nowrap',
         alignItems: 'center',
@@ -98,12 +100,6 @@ const useStyles = createUseStyles({
         'align-items': 'center',
         'gap': '10px',
         'color': props.theme.colors.contentPrimary,
-        '&:hover': {
-            // background: color(props.theme.colors.background)
-            //     .darken(props.themeType === 'light' ? 0.06 : 0.2)
-            //     .rgb()
-            //     .string(),
-        },
     }),
 })
 
@@ -113,68 +109,8 @@ export default function Header() {
     const styles = useStyles({ theme, themeType })
     const headerStyles = useHeaderStyles({ theme })
     const location = useLocation()
-    const errMsgExpireTimeSeconds = 5
-    const lastErrMsgRef = useRef<Record<string, number>>({})
-    const lastLocationPathRef = useRef(location.pathname)
     // eslint-disable-next-line react-hooks/exhaustive-deps
-    const { currentUser, setCurrentUser } = useCurrentUser()
-    const userInfo = useQuery('currentUser', fetchCurrentUser, { enabled: false })
-    // TODO:  refact move to sep file
-    useEffect(() => {
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        if ((axios.interceptors.response as any).handlers.length > 0) {
-            return
-        }
-        axios.interceptors.response.use(
-            (response) => {
-                if (response.headers.authorization) setToken(response.headers.authorization)
-                return response.data?.data ? response.data : response
-            },
-            (error) => {
-                const errMsg = getErrMsg(error)
-                if (error.response?.status === 403 && error.config.method === 'get') {
-                    const search = qs.parse(location.search, { ignoreQueryPrefix: true })
-                    let { redirect } = search
-                    if (redirect && typeof redirect === 'string') {
-                        redirect = decodeURI(redirect)
-                    } else if (['/login', '/logout'].indexOf(location.pathname) < 0) {
-                        redirect = `${location.pathname}${location.search}`
-                    } else {
-                        redirect = '/'
-                    }
-                    if (location.pathname !== '/login' && location.pathname !== '/login/') {
-                        window.location.href = `${window.location.protocol}//${
-                            window.location.host
-                        }/login?redirect=${encodeURIComponent(redirect)}`
-                    }
-                } else if (Date.now() - (lastErrMsgRef.current[errMsg] || 0) > errMsgExpireTimeSeconds * 1000) {
-                    toaster.negative(errMsg, { autoHideDuration: (errMsgExpireTimeSeconds + 1) * 1000 })
-                    lastErrMsgRef.current[errMsg] = Date.now()
-                }
-                return Promise.reject(error)
-            }
-        )
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [])
-
-    useEffect(() => {
-        if (userInfo.isSuccess) {
-            setCurrentUser(userInfo.data)
-        }
-    }, [userInfo.data, userInfo.isSuccess, setCurrentUser])
-
-    useEffect(() => {
-        if (location.pathname !== '/login' && location.pathname !== '/login/' && !currentUser) {
-            userInfo.refetch()
-        }
-    }, [userInfo, location.pathname, currentUser])
-
-    useEffect(() => {
-        if (lastLocationPathRef.current !== location.pathname) {
-            lastErrMsgRef.current = {}
-        }
-        lastLocationPathRef.current = location.pathname
-    }, [location.pathname])
+    const { currentUser } = useCurrentUser()
 
     const [t] = useTranslation()
