+++ console/src/theme/index.ts
@@ -15,38 +15,11 @@ export interface IComposedComponentProps {
     className?: string
 }
 
-// TODO: type support
-// export interface IColors extends Partial<Colors>  {
-//     brand1: string
-//     brand2: string
-//     brandBackground: string
-// }
-
 export const colors = {
-    brand1: '#273343',
-    brand2: '#0C1B3E',
-    brandBackground1: '#F1F6FF',
-    brandBackground2: '#F3F6F9',
-    brandBackground3: '#E7EBF0',
-    brandBackground4: '#E0E3E7',
-    brandFontRmphasis: '#1A2027',
-    brandFontRegular: '#2D3843',
-    brandFontCaption: '#3E5060',
-    brandFontDisable: '#BEC2C2',
-    brandIndicatorAction1: '#A0AAB4',
-    brandIndicatorRegular: '#BFC7CF',
-    brandIndicatorDisabled: color('#000000').alpha(2.5).toString(),
-    brandIndicatorCaption: color('#000000').alpha(4.5).toString(),
-    brandIndicatorSuccess: '#34B576',
-    brandIndicatorError: '#E82037',
-    brandIndicatorWarning: '#FFCD00',
-    brandIndicatorSuspend: '#722ED1',
-    brandIndicatorCompleted: '#91D5FF',
-    brandLink: '#009BDE',
-    //
     brandPrimary: '#2B65D9',
     brandPrimaryHover: color('#2B65D9').alpha(0.5).toString(),
     brandFontPrimary: '#02102B',
+    brandBackgroundPrimar: '#EBF1FF',
     brandBgNav: '#122A59',
     brandBgNavTitle: '#1D3973',
     brandBgNavBorder: '#1D3973',
@@ -54,8 +27,11 @@ export const colors = {
     brandBgNavFontGray: color('#FFFFFF').alpha(0.6).toString(),
     brandBgContent: '#EBF1FF',
     brandBgUser: '#D0DDF7',
+    brandBgSecondory: '#D0DDF7',
+    brandBgSecondory4: color('#D0DDF7').alpha(0.4).toString(),
     brandBgUserFont: '#02102B',
     brandTableHeaderBackground: '#F3F5F9',
+    brandLink: '#2B65D9',
 }
 
 const primitives = {
@@ -80,8 +56,7 @@ const overrides = {
         colors: {
             buttonPrimaryFill: colors.brandPrimary,
             // ----------- custom -----------
-            brandRootBackground: colors.brandBackground1,
-            brandHeaderBackground: colors.brand2,
+            brandRootBackground: colors.brandBackgroundPrimar,
             ...colors,
         },
         typography: {},
